refactor(enrichers): extract shared helpers to eliminate duplicate code
Consolidate duplicated logic across Azure enrichers into shared helpers:
- probeSCMSite, checkIPRestrictionsCommand, checkEasyAuthCommand for
  App Service and Function App enrichers (~330 lines removed)
- buildFirewallRulesCommand with callback pattern for SQL, PostgreSQL,
  and MySQL firewall rule fetchers (shared output struct + JSON marshal)
- enrichEventGridPOSTEndpoint for Event Grid Topic and Domain enrichers
- derefString/derefInt32 helpers for safe pointer dereferences

Also fix error wrapping: %v → %w in container_registry.go.
