+++ pkg/modules/azure/enrichers/sql_server.go
@@ -2,9 +2,7 @@ package enrichers
 
 import (
 	"context"
-	"encoding/json"
 	"fmt"
-	"strings"
 	"time"
 
 	"github.com/Azure/azure-sdk-for-go/sdk/resourcemanager/sql/armsql/v2"
@@ -41,11 +39,12 @@ func enrichSQLServer(cfg plugin.AzureEnricherConfig, result *templates.ARGQueryR
 
 func getSQLFirewallRulesCommand(cfg plugin.AzureEnricherConfig, subscriptionID, resourceGroup, serverName string) plugin.AzureEnrichmentCommand {
 	azCommand := fmt.Sprintf("az sql server firewall-rule list --resource-group %s --server %s", resourceGroup, serverName)
+	description := "Retrieve SQL Server firewall rules"
 
 	if serverName == "" || subscriptionID == "" || resourceGroup == "" {
 		return plugin.AzureEnrichmentCommand{
 			Command:      azCommand,
-			Description:  "Retrieve SQL Server firewall rules",
+			Description:  description,
 			ActualOutput: "Error: Server name, subscription ID, or resource group is missing",
 			ExitCode:     1,
 		}
@@ -56,86 +55,33 @@ func getSQLFirewallRulesCommand(cfg plugin.AzureEnricherConfig, subscriptionID, 
 		ctx = context.Background()
 	}
 
-	clientFactory, err := armsql.NewClientFactory(subscriptionID, cfg.Credential, nil)
-	if err != nil {
-		return plugin.AzureEnrichmentCommand{
-			Command:      azCommand,
-			Description:  "Retrieve SQL Server firewall rules (SDK failed)",
-			ActualOutput: fmt.Sprintf("SDK retrieval failed: %s", err.Error()),
-			Error:        err.Error(),
-			ExitCode:     1,
-		}
-	}
-
-	firewallClient := clientFactory.NewFirewallRulesClient()
-	pager := firewallClient.NewListByServerPager(resourceGroup, serverName, nil)
-
-	type firewallRuleOutput struct {
-		EndIPAddress   string `json:"endIpAddress"`
-		ID             string `json:"id"`
-		Name           string `json:"name"`
-		ResourceGroup  string `json:"resourceGroup"`
-		StartIPAddress string `json:"startIpAddress"`
-		Type           string `json:"type"`
-	}
-
-	var outputRules []firewallRuleOutput
-	for pager.More() {
-		page, err := pager.NextPage(ctx)
+	return buildFirewallRulesCommand(azCommand, description, func() ([]firewallRuleOutput, error) {
+		clientFactory, err := armsql.NewClientFactory(subscriptionID, cfg.Credential, nil)
 		if err != nil {
-			return plugin.AzureEnrichmentCommand{
-				Command:      azCommand,
-				Description:  "Retrieve SQL Server firewall rules (SDK failed)",
-				ActualOutput: fmt.Sprintf("SDK retrieval failed: %s", err.Error()),
-				Error:        err.Error(),
-				ExitCode:     1,
-			}
+			return nil, err
 		}
-		for _, rule := range page.Value {
-			if rule == nil {
-				continue
-			}
-			o := firewallRuleOutput{Type: "Microsoft.Sql/servers/firewallRules"}
-			if rule.Name != nil {
-				o.Name = *rule.Name
-			}
-			if rule.ID != nil {
-				o.ID = *rule.ID
-				parts := strings.Split(o.ID, "/")
-				for i, part := range parts {
-					if part == "resourceGroups" && i+1 < len(parts) {
-						o.ResourceGroup = parts[i+1]
-						break
-					}
-				}
-			}
-			if rule.Properties != nil {
-				if rule.Properties.StartIPAddress != nil {
-					o.StartIPAddress = *rule.Properties.StartIPAddress
-				}
-				if rule.Properties.EndIPAddress != nil {
-					o.EndIPAddress = *rule.Properties.EndIPAddress
-				}
-			}
-			outputRules = append(outputRules, o)
-		}
-	}
 
-	output := "[]"
-	if len(outputRules) > 0 {
-		b, err := json.MarshalIndent(outputRules, "", "  ")
-		if err != nil {
-			output = fmt.Sprintf("Error formatting output: %s", err.Error())
-		} else {
-			output = string(b)
+		pager := clientFactory.NewFirewallRulesClient().NewListByServerPager(resourceGroup, serverName, nil)
+		var rules []firewallRuleOutput
+		for pager.More() {
+			page, err := pager.NextPage(ctx)
+			if err != nil {
+				return nil, err
+			}
+			for _, rule := range page.Value {
+				if rule == nil || rule.Properties == nil {
+					continue
+				}
+				rules = append(rules, firewallRuleOutput{
+					Type:           "Microsoft.Sql/servers/firewallRules",
+					Name:           derefString(rule.Name),
+					ID:             derefString(rule.ID),
+					ResourceGroup:  ParseResourceGroup(derefString(rule.ID)),
+					StartIPAddress: derefString(rule.Properties.StartIPAddress),
+					EndIPAddress:   derefString(rule.Properties.EndIPAddress),
+				})
+			}
 		}
-	}
-
-	return plugin.AzureEnrichmentCommand{
-		Command:                   azCommand,
-		Description:               "Retrieve SQL Server firewall rules",
-		ExpectedOutputDescription: "List of firewall rules with names and IP address ranges",
-		ActualOutput:              output,
-		ExitCode:                  0,
-	}
+		return rules, nil
+	})
 }
